requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
python-dotenv
pydantic-settings
kiteconnect
apscheduler
prometheus-client
redis==5.0.1
